pyproject.toml
[project]
name = "crescer-f-server"
version = "0.1.0"
description = ""
authors = [
    {name = "CoffeeAbuser67",email = "Hservers170224@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
